requirements.txt
streamlit
pandas
openpyxl
pdfplumber
python-docx
pytesseract
pillow
requests
python-dotenv
